+++ lib/router.js
@@ -1732,7 +1732,7 @@ router.get('/jijitang/article/:id', lazyloadRouteHandler('./routes/jijitang/arti
 router.get('/jijitang/publication', lazyloadRouteHandler('./routes/jijitang/publication'));
 
 // Grub Street
-router.get('/grubstreet', lazyloadRouteHandler('./routes/grubstreet/index'));
+// router.get('/grubstreet', lazyloadRouteHandler('./routes/grubstreet/index'));
 
 // 漫画堆
 router.get('/manhuadui/manhua/:name/:serial?', lazyloadRouteHandler('./routes/manhuadui/manhua'));
