+++ lib/v2/grubstreet/maintainer.js
@@ -1,0 +1,4 @@
+module.exports = {
+    '': ['loganrockmore'],
+    '/': ['loganrockmore'],
+};
