+++ lib/v2/grubstreet/radar.js
@@ -1,0 +1,13 @@
+module.exports = {
+    'grubstreet.com': {
+        _name: 'Grub Street',
+        '.': [
+            {
+                title: 'Posts',
+                docs: 'https://docs.rsshub.app/routes/new-media#grub-street',
+                source: ['/'],
+                target: '/grubstreet',
+            },
+        ],
+    },
+};
