+++ lib/v2/grubstreet/router.js
@@ -1,0 +1,3 @@
+module.exports = (router) => {
+    router.get('/', require('./index'));
+};
